+++ pages/4_Faturamento_Bruto_de_Eventos.py
@@ -1,6 +1,5 @@
 import streamlit as st
 import pandas as pd
-import datetime
 from workalendar.america import Brazil
 import warnings
 from utils.components import *
@@ -151,33 +150,6 @@ def main():
 			grafico_linhas_faturamento_modelo_evento(df_eventos_filtrados_por_status_e_ano, id_casa)
 			st.markdown("*Por mês de competência do evento.")
 	st.write("")
-
-	# Recebido X Vencimento
-	with st.container(border=True):
-		col1, col2, col3 = st.columns([0.1, 2.6, 0.1], gap="large", vertical_alignment="center")
-		with col2:
-			st.markdown("## Recebido X Vencimento")
-			st.write("")
-			df_parcelas_recebimento = get_parcelas_por_tipo_data(df_parcelas_filtradas_por_status, df_eventos, "Recebimento (Caixa)", ano)
-			df_parcelas_vencimento = get_parcelas_por_tipo_data(df_parcelas_filtradas_por_status, df_eventos, "Vencimento", ano)
-			grafico_barras_vencimento_x_recebimento(df_parcelas_recebimento, df_parcelas_vencimento, id_casa)
-	
-	# Farol de Parcelas Atrasadas
-	with st.container(border=True):
-		col1, col2, col3 = st.columns([0.1, 2.6, 0.1], gap="large", vertical_alignment="center")
-		with col2:
-			st.markdown(f"## Farol de Parcelas Atrasadas")
-			if not id_casa == -1:
-				df_farol = df_filtrar_casa(df_parcelas_filtradas_por_status, id_casa)
-				df_farol = filtra_parcelas_atrasadas(df_farol)
-			else:
-				df_farol = filtra_parcelas_atrasadas(df_parcelas_filtradas_por_status)
-			df_farol = format_columns_brazilian(df_farol, ['Valor_Parcela'])
-			df_farol = df_format_date_columns_brazilian(df_farol, ['Data_Vencimento', 'Data_Recebimento'])
-			df_farol = rename_colunas_parcelas(df_farol)
-			df_farol = df_farol.drop(columns=['ID Casa'], errors='ignore')
-			st.dataframe(df_farol, use_container_width=True, hide_index=True)
-			st.markdown("Todas as parcelas ainda sem pagamento após a Data de Vencimento.")
 
 	
 if __name__ == '__main__':
